requirements.txt
flask==1.1.1
flask-restplus==0.13.0
SQLAlchemy==1.3.13
flask_cors==3.0.8
flask_sqlalchemy==2.4.1
psycopg2==2.8.4
flask_restplus_sqlalchemy==0.0.1
gunicorn==20.0.4
Werkzeug==0.16.1
pyjwt==1.7.1
pycryptodome==3.9.6
Cryptography==2.8
